Feature/add unittests (#12)
* feat: introduce unit test for the ObsCorrelationScatter method

body: This unit test relies on a new method for loading yaml files
which allows the user to create a substution for an environment
variable identified by the yaml tag !ENVVAR .  If the tag !ENVVAR exists
before a yaml key value, and the environment variable specified between
the {} brackets is set, the yaml load routine will substitude whatever
is set in the specified environment variable.  For instance, if one
set an environment variable 'EVA_TESTS_DIR' to the following:
export EVA_TESTS_DIR=/absolute/path/to/eva/install/location
the the yaml load routine will change the following yaml key

  yaml_key: !ENVVAR ${EVA_TESTS_DIR}/data/datafile.nc4

to

  yaml_key: /absolute/path/to/eva/install/location/data/datafile.nc4

* refactor: remove unnecessary logging and associated lines

* fix: update test yaml config to adopt the new format

body: due to a change made to the expected format of the config
yaml format, I had to fix a few yaml keys (i.e. the yaml key
'diagnostics' needed to be changed to 'applications'. I also fixed
a yaml subkey 'diagnostic name' to 'application name'.

* fix: accidentally moved several print statements, moving them back

* fix: making changes to comply with pycodestyle

* refactor: move and modify yaml loading routine

body: This commit is to revert the yaml loading routine to
what it was prior to the unit test addition, however it is moved
to it's own method so that the unit test can substitute it's own
version of 'load_yaml_file'.  The logger is added to the unit test
as well as the new function'load_yaml_file'.

* fix: fixing bad indentation within a helper function