Fix levels bug in mon_data_space.py (#179)
Levels implementation as dataset coordinate in `data/mon_data_space.py`
was not entirely complete. If the input yaml file requested only a
subset of the available levels, as opposed to requesting all levels
specified in the control file, the results were sub-optimal. The fix is
simple enough -- it now follows the same steps as for the Channel
coordinate.

Resolves #178 .