Add Time Series Capability (#187)
## Description

Add capability to read and transform on the fly and assemble dataset
that has a time series dimension.

- Add test for compiuting mean of omb and making time series.
- Add test for plotting bias correction coefficients time series.
- Add var bc data.
- Add reader for JEDI bias and tlapse files.
- Removed prints of datasets from read and transform classes and added
to driver level.

Plots the new tests create shown below. Note that the tests are somewhat
contrived since the data is the same at both time steps. We can try to
improve on that in the future.


![time_series_omb](https://github.com/JCSDA-internal/eva/assets/27729500/e5d04412-6bde-4f7f-9e4c-da42ca6214d6)


![varbc_time_series](https://github.com/JCSDA-internal/eva/assets/27729500/2ee6db9a-c8ae-476c-b63c-ff9e1782584b)

## Dependencies

NONE

## Impact

NONE

---------

Co-authored-by: danholdaway <danholdaway@users.noreply.github.com>
Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>
Co-authored-by: Akira Sewnath <asewnath@users.noreply.github.com>