Fix config error with emcpy_horizontal_line.py Ref (#204)
Fix horizontal line config error. Use the correct config information in
the plot object.

I ran into this issue last week but assumed it had to be an install
issue, not a simple coding error. I'm not sure how/why this worked as
long as it did. Maybe the unrecognized variable in the plot object is ok
in python3 environments that aren't on Rocky8?

Close #203