Fix 4 problems in mon_data_space and map-scatter plots (#206)
This PR fixes these problems:

1. `mon_data_space.py` now adds empty lat and lon values for each level
when the requested data file is not found. A value of `1` was used
heretofore, which somehow worked. It does not work with code built on
Rocky8 though, hence this change.
2. `emcpy_map_scatter.py` now removes 'level' from the plot object --
emcpy doesn't recognize level and fails if it's in the plot object's
configuration. How this ever worked is not clear to me, but it does with
this change.
3. `dynamic_config.py` now has a check on the data before it tries to
sort the data. Without this check when there was no data an exception
resulted.
4. `map_scatter.py` now includes `level` when selecting the specific
data for a plot. Without this change only plots with a single level
worked; all others had a dimension mismatch and resulting exception in
the color bar generation. This one was totally my fault -- I installed a
similar fix in `line_plot.py` months ago and asked myself if I needed to
change any other plots? Nah, don't think so. Ooops.



Close #205