updates to optimize current scripts (#228)
## Description

Updates to scripts to improve efficiency.

Updates in `figure_driver.py` and the `make_figure()` function avoids
some repeated logic, adds some fail safes to make the code more robust
and makes the figure structure a little neater. Some steps have helped
with optimization of code as well. I have noted in the code that there
are areas where parallelization can be utilized given an option in the
configure YAML files, but did not add the infrastructure yet. If this is
something that may be of interest, especially if we move forward with
this work to replace all monitoring efforts, it could be a good option
if we are plotting hundreds of plots.

## Dependencies

If there are PRs that need to be merged before or along with this one,
please list them below
Waiting on the following PRs:
- [x] waiting on [emcpy
#147](https://github.com/NOAA-EMC/emcpy/pull/147)


## Impact

N/A