Separate handling of groups and variables in timeseries (optional) (#240)
⚠️ I messed up and I already pushed my changes to develop, see the last
two commits. It is a minor change and hopefully won't impact others but
I learned my lesson. It is backward compatible.

This simple change allows you to use either the original configuration
format or the new one with separate groups and variables. Also adds an
example for insitu argo data.

Reduces redundancy IMO, you no longer need to repeat group names in your
configuration.

With the new approach it is possible to define `groups` and `variables`
separately, similar to how it is in other classes.
```yaml
time_series:

  - begin_date: ...
    final_date: ...
    interval: PT6H

    collection: experiment
    groups:
      - omanPassedQc
      - ombgPassedQc
    variables:
      - salinity
      - waterTemperature
    aggregation_methods:
      - mean
      - count
      - std
    dimension:  Location
```

closes #231 and #239