Generate filenames from a template (#244)
## Description

Within a given eva yaml file, this update allows a list of yaml files to
be created given a template, start/end date, interval time, and datetime
format. The purpose of this is to process multiple files without having
to manually add all the filenames in the yaml files.

## Dependencies

If there are PRs that need to be merged before or along with this one,
please list them below
Waiting on the following PRs:
- [ ] waiting on [eva
#236](https://github.com/JCSDA-internal/eva/pull/236)

## Impact

None.